move ipfs experiment classes to their own files (#391)
