+++ UnityProject/Assets/OpenMined/Network/Servers/Ipfs/IpfsExperiment.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+using System.Collections;
+
+namespace OpenMined.Network.Servers.Ipfs
+{
+    public class IpfsExperiment
+    {
+        public string[] jobs;
+
+        public IpfsExperiment(string[] jobs)
+        {
+            this.jobs = jobs;
+        }
+    }
+}
