+++ UnityProject/Assets/OpenMined/Network/Servers/Ipfs/IpfsJobConfig.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+using System.Collections;
+using Newtonsoft.Json.Linq;
+
+namespace OpenMined.Network.Servers.Ipfs
+{
+    public class IpfsJobConfig
+    {
+        [SerializeField] public float lr;
+
+        public IpfsJobConfig(float lr)
+        {
+            this.lr = lr;
+        }
+    }
+}
